requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
tqdm>=4.65.0
networkx>=3.0
<<<<<<< Updated upstream
tiktoken>=0.5.0
numpy>=1.24.0
scipy>=1.10.0 
=======

# Vector Database
faiss-cpu>=1.7.0  # Use faiss-gpu for GPU support

# API & Web
openai>=0.27.0
streamlit>=1.20.0
python-dotenv>=1.0.0

# Configuration
PyYAML>=6.0

# Utilities
asyncio>=3.4.3
aiohttp>=3.8.0

# Development (Optional)
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0

# Flask dependencies (新增)
flask>=3.0.0
flask-cors>=4.0.0 
>>>>>>> Stashed changes
